modify: move modals from components to components/modals
