+++ src/modules/dashboard/nutritionist/workshop/CourseStartModal.tsx
@@ -1,4 +1,4 @@
-import MiniModal from "@/common/components/MiniModal";
+import MiniModal from "@/common/components/modals/MiniModal";
 import { showModal } from "@/common/redux/features/showModal";
 import {
   useCourseGetTimeApiQuery,
